requirements.txt
pandas
numpy==1.26.4
requests==2.27.1
faiss-cpu==1.8.0
matplotlib==3.7.1
streamlit==1.34.0
pillow
torch
torchvision
streamlit-cropper
opencv-contrib-python 